Refactor: share actor between SAC and TQC
